+++ src/components/FullscreenPlayer.tsx
@@ -20,201 +20,9 @@ import PlaybackScheduleBadge from './PlaybackScheduleBadge';
 import { usePlaybackScheduleRemaining } from '../utils/playbackScheduleFormat';
 import { FsLyricsApple } from './fullscreenPlayer/FsLyricsApple';
 import { FsLyricsRail } from './fullscreenPlayer/FsLyricsRail';
-
-function formatTime(seconds: number): string {
-  if (!seconds || isNaN(seconds)) return '0:00';
-  const m = Math.floor(seconds / 60);
-  const s = Math.floor(seconds % 60);
-  return `${m}:${s.toString().padStart(2, '0')}`;
-}
-
-// ─── Album art box — crossfades layers so old art stays visible while new loads ─
-// Uses 300px thumbnails (portrait fallback uses 500px separately).
-//
-// Why onLoad instead of new Image() preload:
-//   React batches setLayers(add invisible) + rAF setLayers(make visible) into one
-//   commit, so the browser never sees opacity:0 and the CSS transition never fires.
-//   Using the DOM img's own onLoad guarantees the element was painted at opacity:0
-//   before we flip it to 1.
-const FsArt = memo(function FsArt({ fetchUrl, cacheKey }: { fetchUrl: string; cacheKey: string }) {
-  // true = show raw fetchUrl immediately as fallback while blob resolves.
-  // PlayerBar uses 128px; FS player uses 300px — different cache keys, no warm hit.
-  // Showing the URL directly avoids the multi-second blank wait.
-  const blobUrl = useCachedUrl(fetchUrl, cacheKey, true);
-
-  const [layers, setLayers] = useState<Array<{ src: string; id: number; vis: boolean }>>([]);
-  const counter = useRef(0);
-  const cleanupTimer = useRef<ReturnType<typeof setTimeout> | null>(null);
-
-  // Add a new invisible layer whenever the blob URL changes.
-  useEffect(() => {
-    if (!blobUrl) return;
-    const id = ++counter.current;
-    setLayers(prev => [...prev, { src: blobUrl, id, vis: false }]);
-  }, [blobUrl]);
-
-  // Called by the DOM <img> once it has painted at opacity:0 — now safe to transition.
-  // Cancel any pending cleanup timer so a stale setTimeout from a previous layer
-  // cannot remove the layer we are making visible right now.
-  const handleLoad = useCallback((id: number) => {
-    if (cleanupTimer.current) clearTimeout(cleanupTimer.current);
-    setLayers(prev => prev.map(l => ({ ...l, vis: l.id === id })));
-    cleanupTimer.current = setTimeout(() => setLayers(prev => prev.filter(l => l.id === id)), 400);
-  }, []);
-
-  if (layers.length === 0) {
-    return <div className="fs-art fs-art-placeholder"><Music size={40} /></div>;
-  }
-
-  return (
-    <>
-      {layers.map(l => (
-        <img
-          key={l.id}
-          src={l.src}
-          className="fs-art"
-          style={{ opacity: l.vis ? 1 : 0 }}
-          onLoad={() => handleLoad(l.id)}
-          alt=""
-          decoding="async"
-        />
-      ))}
-    </>
-  );
-});
-
-// ─── Artist portrait — right half, crossfades on track change ─────────────────
-const FsPortrait = memo(function FsPortrait({ url }: { url: string }) {
-  const [layers, setLayers] = useState<Array<{ url: string; id: number; visible: boolean }>>(() =>
-    url ? [{ url, id: 0, visible: true }] : []
-  );
-  const counterRef = useRef(1);
-  const cleanupTimer = useRef<ReturnType<typeof setTimeout> | null>(null);
-
-  useEffect(() => {
-    if (!url) return;
-    let cancelled = false;
-    const id = counterRef.current++;
-    const img = new Image();
-    img.onload = img.onerror = () => {
-      if (cancelled) return;
-      setLayers(prev => [...prev, { url, id, visible: false }]);
-      requestAnimationFrame(() => {
-        if (cancelled) return;
-        if (cleanupTimer.current) clearTimeout(cleanupTimer.current);
-        setLayers(prev => prev.map(l => ({ ...l, visible: l.id === id })));
-        cleanupTimer.current = setTimeout(() => {
-          if (!cancelled) setLayers(prev => prev.filter(l => l.id === id));
-        }, 1000);
-      });
-    };
-    img.src = url;
-    return () => { cancelled = true; };
-  }, [url]);
-
-  if (layers.length === 0) return null;
-
-  return (
-    <div className="fs-portrait-wrap" aria-hidden="true">
-      {layers.map(layer => (
-        <img
-          key={layer.id}
-          src={layer.url}
-          className="fs-portrait"
-          style={{ opacity: layer.visible ? 1 : 0 }}
-          decoding="async"
-          loading="eager"
-          alt=""
-        />
-      ))}
-    </div>
-  );
-});
-
-// ─── Full-width seekbar — imperative DOM updates, zero React re-renders on tick ─
-const FsSeekbar = memo(function FsSeekbar({ duration }: { duration: number }) {
-  const seek        = usePlayerStore(s => s.seek);
-  const timeRef     = useRef<HTMLSpanElement>(null);
-  const playedRef   = useRef<HTMLDivElement>(null);
-  const bufRef      = useRef<HTMLDivElement>(null);
-  const inputRef    = useRef<HTMLInputElement>(null);
-  const isDraggingRef = useRef(false);
-  const pendingSeekRef = useRef<number | null>(null);
-
-  const previewSeek = useCallback((progress: number) => {
-    const s = usePlayerStore.getState();
-    const p = Math.max(0, Math.min(1, progress));
-    pendingSeekRef.current = p;
-    if (timeRef.current) {
-      const previewTime = duration > 0 ? p * duration : s.currentTime;
-      timeRef.current.textContent = formatTime(previewTime);
-    }
-    if (playedRef.current) playedRef.current.style.width = `${p * 100}%`;
-    if (bufRef.current) bufRef.current.style.width = `${Math.max(p * 100, s.buffered * 100)}%`;
-    if (inputRef.current) inputRef.current.value = String(p);
-  }, [duration]);
-
-  const commitSeek = useCallback(() => {
-    const pending = pendingSeekRef.current;
-    if (pending === null) return;
-    pendingSeekRef.current = null;
-    seek(pending);
-  }, [seek]);
-
-  useEffect(() => {
-    const s = getPlaybackProgressSnapshot();
-    const pct = s.progress * 100;
-    if (timeRef.current)   timeRef.current.textContent  = formatTime(s.currentTime);
-    if (playedRef.current) playedRef.current.style.width = `${pct}%`;
-    if (bufRef.current)    bufRef.current.style.width    = `${Math.max(pct, s.buffered * 100)}%`;
-    if (inputRef.current)  inputRef.current.value        = String(s.progress);
-
-    return subscribePlaybackProgress(state => {
-      if (isDraggingRef.current) return;
-      const p = state.progress * 100;
-      if (timeRef.current)   timeRef.current.textContent  = formatTime(state.currentTime);
-      if (playedRef.current) playedRef.current.style.width = `${p}%`;
-      if (bufRef.current)    bufRef.current.style.width    = `${Math.max(p, state.buffered * 100)}%`;
-      if (inputRef.current)  inputRef.current.value        = String(state.progress);
-    });
-  }, []);
-
-  const handleSeek = useCallback(
-    (e: React.ChangeEvent<HTMLInputElement>) => {
-      previewSeek(parseFloat(e.target.value));
-    },
-    [previewSeek]
-  );
-
-  return (
-    <div className="fs-seekbar-wrap">
-      <div className="fs-seekbar-times">
-        <span ref={timeRef} />
-        <span>{formatTime(duration)}</span>
-      </div>
-      <div className="fs-seekbar">
-        <div className="fs-seekbar-bg" />
-        <div className="fs-seekbar-buf" ref={bufRef} />
-        <div className="fs-seekbar-played" ref={playedRef} />
-        <input
-          ref={inputRef}
-          type="range" min={0} max={1} step={0.001}
-          defaultValue={0}
-          onChange={handleSeek}
-          onMouseDown={() => { isDraggingRef.current = true; }}
-          onMouseUp={() => { isDraggingRef.current = false; commitSeek(); }}
-          onTouchStart={() => { isDraggingRef.current = true; }}
-          onTouchEnd={() => { isDraggingRef.current = false; commitSeek(); }}
-          onPointerDown={() => { isDraggingRef.current = true; }}
-          onPointerUp={() => { isDraggingRef.current = false; commitSeek(); }}
-          onKeyUp={commitSeek}
-          onBlur={() => { isDraggingRef.current = false; commitSeek(); }}
-          aria-label="seek"
-        />
-      </div>
-    </div>
-  );
-});
+import { FsArt } from './fullscreenPlayer/FsArt';
+import { FsPortrait } from './fullscreenPlayer/FsPortrait';
+import { FsSeekbar } from './fullscreenPlayer/FsSeekbar';
 
 // ─── Lyrics settings popover — shown above the mic button ────────────────────
 interface FsLyricsMenuProps {
